+++ raytracer/scene.h
@@ -9,9 +9,9 @@
 #include "sphere.h"
 
 typedef struct viewport {
-    int w;
-    int h;
-    int d;
+    float w;
+    float h;
+    float d;
 } ViewPort;
 
 typedef struct scene {
@@ -22,7 +22,7 @@ typedef struct scene {
     LightArray lights;
 } Scene;
 
-ViewPort viewport_new(int w, int h, int d);
+ViewPort viewport_new(float w, float h, float d);
 Vec3 viewport_coords(ViewPort *self, int x, int y, Canvas *canvas);
 
 Scene scene_new(Camera cam, ViewPort viewport, Color bg);
